修改了一下伪彩色之前的像素分层代码，解决了之前colorbar高值大于最大数值后，全部数值为0以及最小层每次都是0的情况

+++ ImageProcess/imageprocess.cpp
@@ -901,8 +901,8 @@ void ImageProcess::sliceInputLuminescneceImage(ImageType::Pointer inputImage,Pix
 		++tempIt;
 	}
 	int sliceMin,sliceMax;
-	sliceMax=(int)((maxImageValue-LValue)/(HValue-LValue)*255);
-	sliceMin=(int)((minImageValue-LValue)/(HValue-LValue)*255);
+	sliceMax=(int)(((float)maxImageValue-LValue)/((float)HValue-LValue)*255);
+	sliceMin=(int)(((float)minImageValue-LValue)/((float)HValue-LValue)*255);
 	if(sliceMin==0)
 		sliceMin=1;//方便进行接下来的矫正
 	if(sliceMax<2)
